# swagger.yaml
openapi: 3.0.3
info:
  title: Pipu Pro
  description: >-
    # Pipu Pro  api documentation
  version: 1.0.0
  contact: { email: ammuftau74@gmail.com }
servers:
  - url: http://localhost:9000/api/v1
    description: This is my local Server
  - url: https://pipu-pro.onrender.com/api/v1
    description: This is the original api URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /meetings/{meetingId}:
    get:
      tags:
        - 🔄 General
      summary: Get meeting by Id
      description: >-
        The API endpoint allows you to retrieve the meeting by id
      operationId: getMeetingById
      parameters:
        - name: meetingId
          in: path
          required: true
          description: The meeting ID
      responses:
        '200':
          description: Get meeting by id
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '443'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:47 GMT
            ETag:
              schema:
                type: string
                example: W/"1bb-YVaffq7t8o3of0yDl9g26tprr/8"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '493'
            RateLimit-Reset:
              schema:
                type: string
                example: '631'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
    patch:
      tags:
        - 🔄 General
      summary: Updated meeting
      description: >-
        The API endpoint allows you update a meeting.
      operationId: UpdateMeeting
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: monday meeting
                description:
                  type: string
                  example: 'general meeting'
                date:
                  type: string
                  example: '2025-07-28T06:42:26.247Z'
                maxUsers:
                  type: number
                  example: 20
                duration:
                  type: number
                  example: 200000000
      responses:
        '201':
          description: Meeting Updated
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '500'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:09 GMT
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '495'
            RateLimit-Reset:
              schema:
                type: string
                example: '772'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: >-
                      Meeting updated successfully.
  /meetings/code:
    get:
      tags:
        - 🔄 General
      summary: Get meeting by code
      description: >-
        The API endpoint allows you to retrieve the meeting by code
      operationId: getMeetingBycode
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  example: J1L0jAj
      responses:
        '200':
          description: Get meeting by code
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '443'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:47 GMT
            ETag:
              schema:
                type: string
                example: W/"1bb-YVaffq7t8o3of0yDl9g26tprr/8"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '493'
            RateLimit-Reset:
              schema:
                type: string
                example: '631'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
  /meetings/join:
    post:
      tags:
        - 🔄 General
      summary: Join meeting
      description: >-
        The API endpoint allows you to join a meeting.
      operationId: joinMeeting
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                meetingId:
                  type: string
                  example: koinni990
                isAudioEnabled:
                  type: boolean
                  example: true
                isVideoEnabled:
                  type: boolean
                  example: true
      responses:
        '201':
          description: Join meeting
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '500'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:09 GMT
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '495'
            RateLimit-Reset:
              schema:
                type: string
                example: '772'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: >-
                      Meeting joined successfully.
...
  /meetings/leave:
    post:
      tags:
        - 🔄 General
      summary: Leave meeting
      description: >-
        The API endpoint allows you to leave a meeting.
      operationId: LeaveMeeting
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                meetingId:
                  type: string
                  example: 683c417a844b0fd86c3d2829
      responses:
        '201':
          description: Leave meeting
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '500'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:09 GMT
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '495'
            RateLimit-Reset:
              schema:
                type: string
                example: '772'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: >-
                      Leave meeting successfully.
...
  /messages:
    post:
      tags:
        - 🔄 General
      summary: send message
      description: >-
        The API endpoint allows you to send chat message.
      operationId: sendMessage
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                chatId:
                  type: string
                  example: 683c89c87f16c6acb1ca3d84
                content:
                  type: string
                  example: Hello, how are you?
      responses:
        '201':
          description: New message sent
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '500'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:09 GMT
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '495'
            RateLimit-Reset:
              schema:
                type: string
                example: '772'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: >-
                      Message sent successfully.
...